+++ football-microservice/src/main/java/com/barath/football/app/entity/Team.java
@@ -14,13 +14,13 @@ public class Team {
 
     @Id
     @Field
-    private int teamId;
+    private Long teamId;
 
     @Field
     private String teamName;
 
     @PersistenceConstructor
-    public Team(int teamId, String teamName) {
+    public Team(Long teamId, String teamName) {
         this.teamId = teamId;
         this.teamName = teamName;
     }
@@ -28,11 +28,11 @@ public class Team {
     public Team() {
     }
 
-    public int getTeamId() {
+    public Long getTeamId() {
         return teamId;
     }
 
-    public void setTeamId(int teamId) {
+    public void setTeamId(Long teamId) {
         this.teamId = teamId;
     }
